# handler/updateOpening.go
package handler

import (
	"net/http" //defini os códigos de status para as requisiçõs HTTP

	"github.com/gin-gonic/gin"
	"github.com/itslucasmiranda/gojob/schemas"
)

// @BasePath /api/v1

// @Summary Update opening
// @Description Update a job opening
// @Tags Openings
// @Accept json
// @Produce json
// @Param id query string true "Opening Identification"
// @Param opening body UpdateOpeningRequest true "Opening data to Update"
// @Success 200 {object} UpdateOpeningResponse
// @Failure 400 {object} ErrorResponse
// @Failure 404 {object} ErrorResponse
// @Failure 500 {object} ErrorResponse
// @Router /opening [put]
func UpdateOpeningHandler(ctx *gin.Context) { //contexto da requisição
	request := UpdateOpeningRequest{} //cria uma instância da estrutura que representa os dados recebidos no corpo da requisição para atualizar a vaga

	ctx.BindJSON(&request) //converte os dados recebidos para a estrutura JSON recebido para a estrutura UpdateOpeningRequest

	if err := request.Validate(); err != nil {
		logger.Errorf("validation error: %v", err.Error()) //verifica se os campos obrigatórios estão presentes e sõa válidos
		sendError(ctx, http.StatusBadRequest, err.Error())
		return
	}

	id := ctx.Query("id") //obtêm o parametro id da URL
	if id == "" {
		sendError(ctx, http.StatusBadRequest, errParamIsRequired("id", "queryParameter").Error())
		return
	}
	opening := schemas.Opening{} //cria uma variável para armazenar os dados das vagas

	if err := db.First(&opening, id).Error; err != nil { //utiliza o GORM para buscar vagas no banco pelo id
		sendError(ctx, http.StatusNotFound, "opening not found")
		return
	}
	// Update opening
	if request.Role != "" {
		opening.Role = request.Role
	}

	if request.Company != "" {
		opening.Company = request.Company
	}

	if request.Location != "" {
		opening.Location = request.Location
	}

	if request.Remote != nil {
		opening.Remote = *request.Remote
	}

	if request.Link != "" {
		opening.Link = request.Link
	}

	if request.Salary > 0 {
		opening.Salary = request.Salary
	}
	// Save opening
	if err := db.Save(&opening).Error; err != nil {
		logger.Errorf("error updating opening: %v", err.Error())
		sendError(ctx, http.StatusInternalServerError, "error updating opening")
		return
	}
	sendSuccess(ctx, "update-opening", opening)
}